+++ src/pages/FlowRun/FlowRun.vue
@@ -16,6 +16,7 @@ import TaskRunTableTile from '@/pages/FlowRun/TaskRunTable-Tile'
 import TileLayout from '@/layouts/TileLayout'
 import TileLayoutFull from '@/layouts/TileLayout-Full'
 import { parser } from '@/utils/markdownParser'
+import { FINISHED_STATES } from '@/utils/states'
 
 export default {
   metaInfo() {
@@ -53,7 +54,9 @@ export default {
   },
   data() {
     return {
+      error: false,
       flowRunNameLoading: false,
+      loadingKey: 0,
       tab: this.getTab(),
       tabs: [
         {
@@ -93,6 +96,9 @@ export default {
     },
     flowRunId() {
       return this.$route.params.id
+    },
+    loading() {
+      return this.loadingKey > 0
     }
   },
   watch: {
@@ -118,9 +124,6 @@ export default {
         default:
           break
       }
-    },
-    flowRun(val) {
-      if (val === 'not-found') this.$router.push({ name: 'not-found' })
     }
   },
   beforeMount() {
@@ -187,19 +190,46 @@ export default {
           id: this.flowRunId
         }
       },
+      loadingKey: 'loadingKey',
       error(error) {
-        if (error.toString().includes('invalid input'))
-          this.$router.push({ name: 'not-found' })
+        if (error.toString().includes('invalid input')) {
+          this.error = true
+        }
       },
       pollInterval: 5000,
-      update: data => data.flow_run_by_pk || 'not-found'
+      update(data) {
+        if (!data) return
+
+        const flowRun = data.flow_run_by_pk
+        if (
+          FINISHED_STATES.includes(flowRun.state) &&
+          flowRun.state !== 'Scheduled'
+        ) {
+          this.$apollo.queries.flowRun.stopPolling()
+
+          this.$apollo.queries.flowRun.startPolling(60000)
+        }
+        return flowRun
+      }
     }
   }
 }
 </script>
 
 <template>
-  <v-sheet v-if="flowRun" color="appBackground">
+  <v-sheet
+    v-if="error && !flowRun"
+    color="appBackground"
+    class="position-relative"
+  >
+    <div
+      class="text-center position-absolute center-absolute text-h4 grey--text text--darken-2"
+      style="z-index: 1;"
+    >
+      Sorry, we weren't able to find this run.
+    </div>
+  </v-sheet>
+  <v-sheet v-else-if="flowRun" color="appBackground">
     <SubPageNav icon="pi-flow-run" page-type="Flow Run">
       <span
         slot="page-title"
@@ -275,10 +305,13 @@ export default {
           <TimelineTile
             v-if="flowId"
             slot="row-0"
-            condensed
             :flow-id="flowId"
             :flow-run-id="flowRunId"
-            :flow-run="flowRun"
+            :flow-run-end-time="flowRun.end_time"
+            :flow-run-scheduled-start-time="flowRun.scheduled_start_time"
+            :flow-run-start-time="flowRun.start_time"
+            :flow-run-state="flowRun.state"
+            :flow-run-states="flowRun.states"
           />
 
           <DetailsTile slot="row-2-col-1-row-1-tile-1" :flow-run="flowRun" />
@@ -359,6 +392,14 @@ export default {
     </v-bottom-navigation>
   </v-sheet>
 </template>
+
+<style lang="scss" scoped>
+.center-absolute {
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+}
+</style>
 
 <style lang="scss">
 /* stylelint-disable */