+++ src/pages/TeamSettings/Account/Billing.vue
@@ -92,7 +92,7 @@ export default {
       <transition name="fade-expand" mode="out-in">
         <div v-if="!isSelfServe && !loading" class="text-subtitle-1">
           To update your payment details, please
-          <a href="https://www.prefect.io/get-prefect#contact" target="_blank">
+          <a href="https://www.prefect.io/pricing#contact" target="_blank">
             contact our sales team
           </a>
         </div>
@@ -233,7 +233,7 @@ export default {
           Sorry, we're having trouble displaying your payment method. For
           questions, please contact us via email at
           <span class="primary--text">support@prefect.io</span> or on our
-          <router-link :to="{ name: 'help' }">support page</router-link>
+          <router-link :to="'/help'">support page</router-link>
         </div>
       </transition>
     </v-card-text>