+++ tests/unit/store/auth.legacy.spec.js
@@ -8,13 +8,14 @@ import {
   getUser,
   idToken,
   accessToken,
-  setTokens,
-  signOut,
-  OktaAuth
+  signOut
 } from '@okta/okta-auth-js'
 jest.mock('@okta/okta-auth-js')
 
-import auth from '@/store/auth'
+import auth_base from '@/store/auth'
+import auth_actions from '@/store/auth/legacy-actions'
+
+const auth = { ...auth_base, ...auth_actions }
 
 // Mock the setauthUser method
 // which is a root module we don't want to
@@ -24,7 +25,6 @@ auth.mutations['user/setOktaUser'] = jest.fn()
 import {
   MOCK_AUTHORIZATION_TOKEN,
   MOCK_ID_TOKEN,
-  MOCK_ID_TOKEN_2,
   MOCK_REFRESH_TOKEN,
   MOCK_PREFECT_AUTH_PAYLOAD
 } from './mockTokens'
@@ -66,6 +66,7 @@ describe('auth Vuex Module', () => {
       isLoggingInUser: false,
       isRefreshingAuthentication: false,
       isRefreshingAuthorization: false,
+      legacy: false,
       redirectRoute: null,
       refreshToken: MOCK_REFRESH_TOKEN,
       refreshTokenExpiry: jwt_decode(MOCK_REFRESH_TOKEN).exp * 1000,
@@ -89,6 +90,7 @@ describe('auth Vuex Module', () => {
       isLoggingInUser: false,
       isRefreshingAuthentication: false,
       isRefreshingAuthorization: false,
+      legacy: false,
       redirectRoute: null,
       refreshToken: null,
       refreshTokenExpiry: null,
@@ -100,7 +102,7 @@ describe('auth Vuex Module', () => {
     it('should be initialized properly without tokens in localStorage ', () => {
       const state = auth.state
 
-      expect(Object.keys(state).length).toBe(16)
+      expect(Object.keys(state).length).toBe(17)
 
       expect(state.user).toBe(null)
       expect(state.authorizationToken).toBe(null)
@@ -211,6 +213,10 @@ describe('auth Vuex Module', () => {
       expect(store.getters.isAuthorizingUser).toBe(
         store.state.isAuthorizingUser
       )
+    })
+
+    it('should return legacy', () => {
+      expect(store.getters.legacy).toBe(store.state.legacy)
     })
 
     it('should return the redirectRoute', () => {